Use hard-coded list of built-in warning categories.
There just turned out to be too much ad-hocery involved in deriving it
from the %warnings::Bits hash keys, due to the fact that
warnings::register modifies this hash, and there is no way to guarantee
access to a pristine copy of the hash. I considered parsing warnings.pm
in App::Warning::Diagnostics when it was loaded, but that felt fragile.
Better to do the ad-hocery once for all.

Made warning_diagnostics() return the individual selected diagnostics in
list context, basically to make counting them easier via the Saturn
operator =()=.